# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/coupons/available": {
      "get": {
        "operationId": "CouponController_getAvailableCoupons",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of available coupons"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all available coupons for current user",
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/{code}/validate": {
      "post": {
        "operationId": "CouponController_validateCoupon",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "description": "Coupon code",
            "schema": {
              "example": "FIRST50",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon validation result"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Validate if coupon is valid for cart",
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/{code}/apply": {
      "post": {
        "operationId": "CouponController_applyCoupon",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "description": "Coupon code",
            "schema": {
              "example": "FIRST50",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon applied successfully"
          },
          "400": {
            "description": "Invalid coupon or validation failed"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Apply coupon to cart/order",
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/my-usage": {
      "get": {
        "operationId": "CouponController_getMyUsage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of coupon usage history"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user coupon usage history",
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/recommend": {
      "post": {
        "operationId": "CouponController_recommendCoupons",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recommended coupons with best coupon and alternatives",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bestCoupon": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "SAVE200"
                        },
                        "potentialSavings": {
                          "type": "number",
                          "example": 200
                        },
                        "coupon": {
                          "type": "object"
                        }
                      }
                    },
                    "alternativeCoupons": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "potentialSavings": {
                            "type": "number"
                          },
                          "coupon": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get recommended coupons based on cart",
        "tags": [
          "coupons"
        ]
      }
    },
...
    "/admin/coupons": {
      "post": {
        "operationId": "AdminController_createCoupon",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Coupon created successfully"
          },
          "400": {
            "description": "Invalid coupon data"
          },
          "409": {
            "description": "Coupon code already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new coupon",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "AdminController_getAllCoupons",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "discountType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userSegment",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of coupons"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all coupons with optional filters",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/coupons/{id}": {
      "get": {
        "operationId": "AdminController_getCoupon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Coupon ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon details"
          },
          "404": {
            "description": "Coupon not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get coupon by ID",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_updateCoupon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Coupon ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon updated successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update coupon details",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteCoupon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Coupon ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon deleted successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete coupon (soft delete)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/coupons/{id}/toggle-status": {
      "patch": {
        "operationId": "AdminController_toggleCouponStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Coupon ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon status toggled"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Toggle coupon active status",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/coupons/{id}/analytics": {
      "get": {
        "operationId": "AdminController_getCouponAnalytics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Coupon ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon analytics data"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get coupon performance analytics",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/reports/coupon-usage": {
      "get": {
        "operationId": "AdminController_getCouponUsageReport",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "couponId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon usage report"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get coupon usage report with date filters",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/reports/top-coupons": {
      "get": {
        "operationId": "AdminController_getTopCoupons",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of top coupons to return",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top coupons list"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get top used coupons",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/reports/revenue-impact": {
      "get": {
        "operationId": "AdminController_getRevenueImpact",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue impact data"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get revenue impact report",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coupon Management System API",
    "description": "API documentation for Coupon Code Management System - Backend Technical Challenge",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CartItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "prod-123",
            "description": "Product ID"
          },
          "category": {
            "type": "string",
            "example": "electronics",
            "description": "Product category"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Product quantity"
          },
          "price": {
            "type": "number",
            "example": 1000,
            "description": "Product price"
          }
        },
        "required": [
          "productId"
        ]
      },
      "ValidateCouponDto": {
        "type": "object",
        "properties": {
          "cartValue": {
            "type": "number",
            "example": 1000,
            "description": "Cart value in currency units"
          },
          "items": {
            "description": "Cart items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemDto"
            }
          },
          "paymentMethod": {
            "type": "string",
            "example": "card",
            "description": "Payment method",
            "default": "card"
          }
        },
        "required": [
          "cartValue"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "RecommendCouponDto": {
        "type": "object",
        "properties": {
          "cartValue": {
            "type": "number",
            "example": 1500,
            "description": "Cart value in currency units"
          },
          "items": {
            "description": "Cart items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemDto"
            }
          },
          "userId": {
            "type": "string",
            "example": "user123",
            "description": "User ID (optional, will use current user if not provided)"
          }
        },
        "required": [
          "cartValue"
        ]
      },
...
      "CreateCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "FIRST50",
            "description": "Unique coupon code"
          },
          "title": {
            "type": "string",
            "example": "50% off on first order",
            "description": "Coupon title"
          },
          "description": {
            "type": "string",
            "example": "Get 50% off on your first order",
            "description": "Coupon description"
          },
          "discountType": {
            "type": "string",
            "enum": [
              "percentage",
              "fixed_amount",
              "free_delivery"
            ],
            "example": "percentage",
            "description": "Type of discount"
          },
          "discountValue": {
            "type": "number",
            "example": 50,
            "description": "Discount value (percentage or fixed amount)",
            "minimum": 0
          },
          "minOrderValue": {
            "type": "number",
            "example": 500,
            "description": "Minimum order value required",
            "minimum": 0
          },
          "maxDiscountCap": {
            "type": "number",
            "example": 200,
            "description": "Maximum discount cap (required for percentage)",
            "minimum": 0
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00Z",
            "description": "Coupon start date"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-12-31T23:59:59Z",
            "description": "Coupon end date"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether coupon is active",
            "default": true
          },
          "totalUsageLimit": {
            "type": "number",
            "example": 1000,
            "description": "Total usage limit for the coupon",
            "minimum": 1
          },
          "perUserLimit": {
            "type": "number",
            "example": 1,
            "description": "Usage limit per user",
            "minimum": 1
          },
          "applicableCategories": {
            "example": [
              "electronics",
              "groceries"
            ],
            "description": "Applicable product categories",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "applicableProducts": {
            "example": [
              "prod-123",
              "prod-456"
            ],
            "description": "Applicable product IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userSegment": {
            "type": "string",
            "enum": [
              "all",
              "new_users",
              "premium_users"
            ],
            "example": "new_users",
            "description": "Target user segment",
            "default": "all"
          },
          "minPurchaseCount": {
            "type": "number",
            "example": 3,
            "description": "Minimum purchase count required",
            "minimum": 0
          },
          "excludedCategories": {
            "example": [
              "clothing"
            ],
            "description": "Excluded product categories",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludedProducts": {
            "example": [
              "prod-789"
            ],
            "description": "Excluded product IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "paymentMethods": {
            "example": [
              "card",
              "upi"
            ],
            "description": "Allowed payment methods",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "title",
          "description",
          "discountType",
          "discountValue",
          "minOrderValue",
          "startDate",
          "endDate"
        ]
      },
      "UpdateCouponDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}